tests/requirements.txt
pytest==7.4.3
boto3==1.34.7
requests==2.31.0
python-dotenv==1.0.0